+++ src/agent/src/confidential_data_hub/image.rs
